+++ NEWS.md
@@ -1,18 +1,27 @@
 # santoku (development version)
 
+## Breaking changes
+
+* The `groups` argument to `chop_evenly()`,  deprecated in 0.4.0, has 
+  been removed.
+* `brk_left()` and `brk_right()`, deprecated in 0.4.0, have been removed.
+* `knife()`, deprecated in 0.4.0, has been removed.
+
+## Other changes
+
 * You can now chop many more types, including `units` from the `units` package,
   `difftime` objects, `package_version` objects, etc. Characters will be
   chopped with a warning.
 * Experimental `lbl_glue()` function using the `glue` package.
 * You can now set `labels = NULL` to return integer codes.
 * Use unicode em-dash and no spaces in `lbl_dash()` and `lbl_discrete()`.
-* The `groups` argument to `chop_evenly()`,  deprecated in 0.4.0, has 
-  been removed.
+
 * Arguments `first` and `last` can now be used in `lbl_intervals()` to override
   the first and last interval labels.
 * Bugfix: `tab()` and friends no longer display an `x` as the variable name.
 * Bugfix: `lbl_endpoint()` was erroring for some types of breaks.
 * `brk_default()` now throws an error if breaks are not sorted.
+
 
 # santoku 0.6.0